# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dryko CRM API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ContactCanal = typeof ContactCanal[keyof typeof ContactCanal];


export const ContactCanal = {
  whatsapp: 'whatsapp',
  email: 'email',
  instagram: 'instagram',
  telefone: 'telefone',
} as const;

export interface Contact {
  id: number;
  nome: string;
  telefone: string;
  /** @nullable */
  email?: string | null;
  canal: ContactCanal;
  tags: string[];
  /** @nullable */
  notas: string | null;
  criadoEm: string;
  atualizadoEm: string;
}

export type ContactInputCanal = typeof ContactInputCanal[keyof typeof ContactInputCanal];


export const ContactInputCanal = {
  whatsapp: 'whatsapp',
  email: 'email',
  instagram: 'instagram',
  telefone: 'telefone',
} as const;

export interface ContactInput {
  /** @minLength 1 */
  nome: string;
  /** @minLength 1 */
  telefone: string;
  /** @nullable */
  email?: string | null;
  canal: ContactInputCanal;
  tags?: string[];
  /** @nullable */
  notas?: string | null;
}

export type ContactUpdateCanal = typeof ContactUpdateCanal[keyof typeof ContactUpdateCanal];


export const ContactUpdateCanal = {
  whatsapp: 'whatsapp',
  email: 'email',
  instagram: 'instagram',
  telefone: 'telefone',
} as const;

export interface ContactUpdate {
  /** @minLength 1 */
  nome?: string;
  /** @minLength 1 */
  telefone?: string;
  /** @nullable */
  email?: string | null;
  canal?: ContactUpdateCanal;
  tags?: string[];
  /** @nullable */
  notas?: string | null;
}

export interface CanalCount {
  canal: string;
  total: number;
}

export interface ContactStats {
  total: number;
  porCanal: CanalCount[];
}

export type ConversationCanal = typeof ConversationCanal[keyof typeof ConversationCanal];


export const ConversationCanal = {
  whatsapp: 'whatsapp',
  email: 'email',
  instagram: 'instagram',
  telefone: 'telefone',
} as const;

export type ConversationDirecao = typeof ConversationDirecao[keyof typeof ConversationDirecao];


export const ConversationDirecao = {
  entrada: 'entrada',
  saida: 'saida',
} as const;

export interface Conversation {
  id: number;
  contatoId: number;
  conteudo: string;
  canal: ConversationCanal;
  direcao: ConversationDirecao;
  criadoEm: string;
}

export type ConversationInputCanal = typeof ConversationInputCanal[keyof typeof ConversationInputCanal];


export const ConversationInputCanal = {
  whatsapp: 'whatsapp',
  email: 'email',
  instagram: 'instagram',
  telefone: 'telefone',
} as const;

export type ConversationInputDirecao = typeof ConversationInputDirecao[keyof typeof ConversationInputDirecao];


export const ConversationInputDirecao = {
  entrada: 'entrada',
  saida: 'saida',
} as const;

export interface ConversationInput {
  /** @minLength 1 */
  conteudo: string;
  canal: ConversationInputCanal;
  direcao: ConversationInputDirecao;
}

export interface ImportInput {
  registros: ContactInput[];
}

export interface ImportResult {
  importados: number;
  erros: string[];
}

export type InboxConversationCanal = typeof InboxConversationCanal[keyof typeof InboxConversationCanal];


export const InboxConversationCanal = {
  whatsapp: 'whatsapp',
  email: 'email',
  instagram: 'instagram',
  telefone: 'telefone',
} as const;

export type InboxConversationStatus = typeof InboxConversationStatus[keyof typeof InboxConversationStatus];


export const InboxConversationStatus = {
  aberta: 'aberta',
  em_andamento: 'em_andamento',
  resolvida: 'resolvida',
} as const;

export interface InboxConversation {
  id: number;
  contatoId: number;
  contatoNome: string;
  contatoTelefone?: string;
  canal: InboxConversationCanal;
  status: InboxConversationStatus;
  /** @nullable */
  atendente?: string | null;
  criadoEm: string;
  /** @nullable */
  encerradoEm?: string | null;
  /** @nullable */
  classificacao?: string | null;
  /** @nullable */
  ultimaMensagem?: string | null;
  totalMensagens: number;
}

export type InboxConversationInputCanal = typeof InboxConversationInputCanal[keyof typeof InboxConversationInputCanal];


export const InboxConversationInputCanal = {
  whatsapp: 'whatsapp',
  email: 'email',
  instagram: 'instagram',
  telefone: 'telefone',
} as const;

export interface InboxConversationInput {
  contatoId: number;
  canal: InboxConversationInputCanal;
  /** @nullable */
  atendente?: string | null;
}

export type InboxConversationUpdateStatus = typeof InboxConversationUpdateStatus[keyof typeof InboxConversationUpdateStatus];


export const InboxConversationUpdateStatus = {
  aberta: 'aberta',
  em_andamento: 'em_andamento',
  resolvida: 'resolvida',
} as const;

export interface InboxConversationUpdate {
  status?: InboxConversationUpdateStatus;
  /** @nullable */
  atendente?: string | null;
  /** @nullable */
  classificacao?: string | null;
}

export interface InboxStats {
  abertas: number;
  emAndamento: number;
  resolvidas: number;
  total: number;
}

export interface Mensagem {
  id: number;
  conversaId: number;
  texto: string;
  autor: string;
  criadoEm: string;
}

export interface MensagemInput {
  /** @minLength 1 */
  texto: string;
  /** @minLength 1 */
  autor: string;
}

export interface RespostaRapida {
  id: number;
  titulo: string;
  texto: string;
}

export interface RespostaRapidaInput {
  /** @minLength 1 */
  titulo: string;
  /** @minLength 1 */
  texto: string;
}

export type ListContactsParams = {
search?: string;
tag?: string;
canal?: ListContactsCanal;
};

export type ListContactsCanal = typeof ListContactsCanal[keyof typeof ListContactsCanal];


export const ListContactsCanal = {
  whatsapp: 'whatsapp',
  email: 'email',
  instagram: 'instagram',
  telefone: 'telefone',
} as const;

export type ListInboxParams = {
status?: ListInboxStatus;
canal?: ListInboxCanal;
};

export type ListInboxStatus = typeof ListInboxStatus[keyof typeof ListInboxStatus];


export const ListInboxStatus = {
  aberta: 'aberta',
  em_andamento: 'em_andamento',
  resolvida: 'resolvida',
} as const;

export type ListInboxCanal = typeof ListInboxCanal[keyof typeof ListInboxCanal];


export const ListInboxCanal = {
  whatsapp: 'whatsapp',
  email: 'email',
  instagram: 'instagram',
  telefone: 'telefone',
} as const;