# lib/api-zod/src/generated/types/inboxConversation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dryko CRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InboxConversationCanal } from './inboxConversationCanal';
import type { InboxConversationStatus } from './inboxConversationStatus';

export interface InboxConversation {
  id: number;
  contatoId: number;
  contatoNome: string;
  contatoTelefone?: string;
  canal: InboxConversationCanal;
  status: InboxConversationStatus;
  /** @nullable */
  atendente?: string | null;
  criadoEm: Date;
  /** @nullable */
  encerradoEm?: string | null;
  /** @nullable */
  classificacao?: string | null;
  /** @nullable */
  ultimaMensagem?: string | null;
  totalMensagens: number;
}