# lib/api-zod/src/generated/types/inboxConversationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dryko CRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InboxConversationInputCanal } from './inboxConversationInputCanal';

export interface InboxConversationInput {
  contatoId: number;
  canal: InboxConversationInputCanal;
  /** @nullable */
  atendente?: string | null;
}