# lib/api-zod/src/generated/types/inboxConversationUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dryko CRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InboxConversationUpdateStatus } from './inboxConversationUpdateStatus';

export interface InboxConversationUpdate {
  status?: InboxConversationUpdateStatus;
  /** @nullable */
  atendente?: string | null;
  /** @nullable */
  classificacao?: string | null;
}