# lib/api-zod/src/generated/types/inboxStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dryko CRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InboxStats {
  abertas: number;
  emAndamento: number;
  resolvidas: number;
  total: number;
}